backend/requirements.txt
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-multipart>=0.0.9,<1.0.0